Extract hold functionality into new file
