+++ ether-wake.c
@@ -74,19 +74,15 @@ static char usage_msg[] =
 
 #include <sys/socket.h>
 
-#include <sys/types.h>
 #include <sys/ioctl.h>
 #include <linux/if.h>
 
 #include <netpacket/packet.h>
 #include <net/ethernet.h>
 #include <netinet/ether.h>
-#include <time.h>
-#include <sys/wait.h>
 
 #include "nfqueue.h"
-
-#define TIMEOUT 10
+#include "ping.h"
 
 u_char outpack[1000];
 int pktsize;
@@ -103,7 +99,6 @@ u_char wol_passwd[6];
 int wol_passwd_sz = 0;
 
 static int hold = 0;
-static char **ping_argv;
 
 static int opt_no_src_addr = 0, opt_broadcast = 0;
 static int opt_nfqueue_num = -1;
@@ -113,7 +108,6 @@ static int get_dest_addr(const char *arg, struct ether_addr *eaddr);
 static int get_fill(unsigned char *pkt, struct ether_addr *eaddr);
 static int get_wol_pw(const char *optarg);
 static int get_nfqueue_num(const char *optarg);
-static void generate_ping_argv(const char *ip_address, const char *ifname);
 
 int main(int argc, char *argv[])
 {
@@ -259,51 +253,6 @@ int main(int argc, char *argv[])
 		printf("Acting on packets in NFQUEUE %d\n", opt_nfqueue_num);
 	return nfqueue_receive(opt_nfqueue_num, send_magic_packet);
 }
-void generate_ping_argv(const char *ip_address, const char *ifname)
-{
-	const char *ping_arguments[] = {
-		"/bin/ping", "-c", "1", "-W", "1", "-I", ifname, ip_address
-	};
-	const int argument_count =
-		sizeof(ping_arguments) / sizeof(ping_arguments[0]);
-	ping_argv = calloc(argument_count + 1, sizeof(char *));
-	if (!ping_argv) {
-		perror("malloc() ping arguments");
-		exit(EXIT_FAILURE);
-	}
-
-	for (int array_index = 0; array_index < argument_count; array_index++) {
-		char *str_cpy = strdup(ping_arguments[array_index]);
-		if (str_cpy) {
-			ping_argv[array_index] = str_cpy;
-		} else {
-			perror("malloc() ping argument");
-			exit(EXIT_FAILURE);
-		}
-	}
-}
-
-static int execute_ping()
-{
-	pid_t ping_pid;
-	int status;
-
-	ping_pid = fork();
-	if (ping_pid == 0) {
-		fclose(stdout);
-		fclose(stdin);
-		// Must include some PATH for busybox on OpenWRT
-		char *envp[] = { "PATH=/bin:/sbin:/usr/bin:/usr/sbin", NULL };
-		execve("/bin/ping", ping_argv, envp);
-
-		perror("ping process execve failed [%m]");
-		return EXIT_FAILURE;
-	}
-
-	waitpid(ping_pid, &status, 0);
-
-	return WEXITSTATUS(status);
-}
 
 static int send_magic_packet()
 {
@@ -340,20 +289,7 @@ static int send_magic_packet()
 
 	// Wait until host is reachable again
 	if (hold) {
-		static time_t last_time = -1;
-		time_t current_time = time(NULL);
-		int recent = current_time != -1 &&
-			     difftime(current_time, last_time) < 60;
-		if (!recent) {
-			int ping_ret = -1;
-			for (int ping_count = 0;
-			     ping_count < TIMEOUT && ping_ret != 0;
-			     ping_count++) {
-				ping_ret = execute_ping();
-			}
-		}
-
-		last_time = current_time;
+		wait_for_online();
 	}
 
 	return 0;
